Arreglé el método de modificación de Días

+++ src/main/java/com/example/demo/persistencia/clases/DAO/GenericDAOHibernateJPA.java
@@ -37,7 +37,7 @@ public class GenericDAOHibernateJPA<T> implements GenericDAO<T> {
     public boolean exist(long id) {
         try {
             // Consulta optimizada con "SELECT 1", que es más eficiente
-            Long count = em.createQuery("SELECT 1 FROM " + entityClass.getName() + " e WHERE e.id = :id", Long.class)
+            Integer count = em.createQuery("SELECT 1 FROM " + entityClass.getName() + " e WHERE e.id = :id", Integer.class)
                            .setParameter("id", id)
                            .getSingleResult();
             return count != null;
